demo/requirements.txt
accelerate
streamlit
diffusers
Pillow
numpy
opencv-python
transformers

